requirements/full.txt
torch
torchvision
numpy
opencv-python
Pillow
tqdm>=4.36.0
matplotlib
plotly
scipy
omegaconf
hydra-core
pytorch-lightning
torchmetrics
jupyter
lxml
rtree
git+https://github.com/mapillary/OpenSfM
httpx
aiolimiter
scikit-learn
